cliconfig: Move HCL 1.0-based language out of the way
We've been continuing to maintain the HCL 1-based CLI configuration
language to keep compatibility with prior versions, but the HCL 1 API is
constraining and it's annoying to have both versions of HCL in Terraform
at once.

This commit doesn't change any user-visible behavior but it does rename
a bunch of things in the cliconfig package to include the word "Legacy" so
that we can gradually build out an HCL 2 based new implementation
alongside and eventually support them both together.

The plan to support them both at once is to also switch over to using
OS-standard configuration file locations (like XDG Base Directory on Linux)
at the same time and treat the new locations as an opt-in for the new
language variant, with both continuing to be supported for the forseeable
future but the legacy one likely becoming increasingly more deprecated as
it ages.
